Save/Load Edits, Inventory/Equipment, and Standarization/OOP.
Fully built out the save and loading of data and data structures to hold said data. Also began adding framework for keeping track of all spawned/saveable entities in the scene. Also fixed a bit of a bug with animating the test pawn. Added a constants file. Added a bunch of basic scripts for various features. Reworked the map/tiling so that tiles are grouped into layers which are then grouped into map(s). Also had the layers moved horizontally instead of stacked so physics doesn't carry across multiple layers.