Edit load save in main menu and early Save Stat work
Edited the 'load scene' prefabs in the main menu to be ready to take information about each save slot and added buttons to them. Also added the scripts for stats on each save slot (currently just time last played) and began implementing the data being saved and loaded when necessary.