Add in functionality for settings menu and exiting game
Add in code to enable functionality for settings menu. Allows user to set volume, resolution, enable/disable fullscreen mode etc.
Add in functionality to allow user to quit game. Not sure what other options should be allowed for this or if it should be its own script to allow accessibility from other scenes.