Added Hunger Timer (#84)
* Added a hunger timer and pawn death.

Added hunger timer which decrements the hunger of all pawns over time. If a pawn reaches zero, they are killed by removing them from availablePawns in the LaborOrderManager and the new PawnList in Pawn.

* Added test for removeSpecificPawn

Added test for new function removeSpecificPawn in LaborOrderManager.

* Cleaning up

reverted temporary changes to MapManager that use to test.