Tile Hover, Work Area Sizing, and Pawn Count (#100)
* Tile Hover, Work Area Sizing, and Pawn Count

Added feature where when a user hovers over a tile, info on that tile is shown (will need to update this once we incorporate whatever version of PlacedObjects). Also added a UI element to show the size of a Work Area when a user is drag and dropping it. Finally, hooked up LaborOrderManager's "PawnCount()" to the Colony Info UI so the correct pawn count is now shown.

* Forgot to save scene

Just forgot to save the scene before pushing. Believe I just attached some scripts and/or hooked up some attachments.

---------

Co-authored-by: Benjamin Weber <weberb1@ad.unlv.edu>